Playerの攻撃力の変動処理

+++ UnityProject/Assets/Boss/Scripts/BossBehavior.cs
@@ -284,7 +284,7 @@ public class BossBehavior : MonoBehaviour, IPlayerDamege
     }
 
     //ƒ_ƒ[ƒW‚ğó‚¯‚é
-    public void _AddDamege(int _Damage)
+    public void _AddDamege(float _Damage)
     {
         Debug.Log("hit!");
         _Health--;
