+++ UnityProject/Assets/Player/PlayerScript/IPlayerDamege.cs
@@ -1,4 +1,4 @@
 public interface IPlayerDamege
 {
-    void _AddDamege(int _Damege);
+    void _AddDamege(float _Damege);
 }
